+++ functions/Connect-TibberWebSocket.ps1
@@ -16,19 +16,6 @@
         https://developer.tibber.com/docs/guides/calling-api
     #>
     param (
-        # Specifies the URI for the request.
-        # Override default using the TIBBER_WSS_URI environment variable.
-        [Parameter(ValueFromPipelineByPropertyName)]
-        [Alias('URL')]
-        [Uri] $URI = $(
-            if ($env:TIBBER_WSS_URI) {
-                $env:TIBBER_WSS_URI
-            }
-            else {
-                'wss://websocket-api.tibber.com/v1-beta/gql/subscriptions'
-            }
-        ),
-
         # Specifies the access token to use for the communication.
         # Override default using the TIBBER_ACCESS_TOKEN environment variable.
         [Parameter(ValueFromPipelineByPropertyName)]
@@ -44,26 +31,37 @@
 
         # Specifies the number of retry attempts if WebSocket initialization fails.
         [ValidateRange(0, [int]::MaxValue)]
-        [Alias('Retries')]
+        [Alias('Retries', 'MaxRetries')]
         [int] $RetryCount = 5,
-
-        # Specifies for how long in seconds we should wait between retries.
-        [ValidateRange(0, [int]::MaxValue)]
-        [Alias('RetryWaitTime', 'WaitTime')]
-        [int] $RetryWaitTimeInSeconds = 5,
 
         # Specifies the time to wait for WebSocket operations, or -1 to wait indefinitely.
         [Parameter(ValueFromPipelineByPropertyName)]
         [ValidateRange(-1, [int]::MaxValue)]
         [Alias('Timeout')]
-        [int] $TimeoutInSeconds = 10
+        [int] $TimeoutInSeconds = 10,
+
+        # Specifies the user agent (appended to the default).
+        # Override default using the TIBBER_USER_AGENT environment variable.
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [string] $UserAgent = $(
+            if ($env:TIBBER_USER_AGENT) {
+                $env:TIBBER_USER_AGENT
+            }
+        )
     )
 
     begin {
+        # Get GraphQL subscriptions endpoint
+        $query = "{ viewer { websocketSubscriptionUrl } }"
+        [Uri] $wssUri = (Invoke-TibberQuery -Query $query -PersonalAccessToken $PersonalAccessToken).viewer.websocketSubscriptionUrl
+
         # Setup request headers
-        $userAgent = "PSTibber/$($MyInvocation.MyCommand.ScriptBlock.Module.Version)"
-        if ($env:TIBBER_USER_AGENT) {
-            $userAgent += " $env:TIBBER_USER_AGENT"
+        $fullUserAgent = "PSTibber/$($MyInvocation.MyCommand.ScriptBlock.Module.Version)"
+        if ($UserAgent) {
+            $fullUserAgent += " $UserAgent"
+        }
+        else {
+            Write-Warning "Missing user agent, please set using '-UserAgent' or '`$env:TIBBER_USER_AGENT'"
         }
     }
 
@@ -79,15 +77,15 @@
             # Setup WebSocket for communication
             $webSocket = New-Object Net.WebSockets.ClientWebSocket
             $webSocket.Options.AddSubProtocol('graphql-transport-ws')
-            $webSocket.Options.SetRequestHeader('User-Agent', $userAgent)
+            $webSocket.Options.SetRequestHeader('User-Agent', $fullUserAgent)
             $cancellationTokenSource = New-Object Threading.CancellationTokenSource
             $cancellationToken = $cancellationTokenSource.Token
             $recvBuffer = New-Object ArraySegment[byte] -ArgumentList @(, $([byte[]] @(, 0) * 16384))
 
             # Connect WebSocket
-            $result = $webSocket.ConnectAsync($URI, $cancellationToken)
+            $result = $webSocket.ConnectAsync($wssUri, $cancellationToken)
             Wait-WebSocketOp -OperationName 'ConnectAsync' -Result $result -TimeoutInSeconds $TimeoutInSeconds
-            Write-Verbose -Message "WebSocket connected to $URI [User agent = $userAgent]"
+            Write-Verbose -Message "WebSocket connected to $wssUri [User agent = $fullUserAgent]"
 
             # Init WebSocket
             $command = @{
@@ -97,7 +95,7 @@
                 }
             } | ConvertTo-Json -Depth 10
             Write-WebSocket -Data $command -WebSocket $webSocket -CancellationToken $cancellationToken -TimeoutInSeconds $TimeoutInSeconds
-            Write-Verbose -Message "Init message sent to: $URI [connection_init]"
+            Write-Verbose -Message "Init message sent to: $wssUri [connection_init]"
 
             # WebSocket init acknowledgement
             $result = $webSocket.ReceiveAsync($recvBuffer, $cancellationToken)
@@ -106,8 +104,9 @@
             Write-Debug -Message ($webSocket | Select-Object * | Out-String)
             if ($result.Result.CloseStatus) {
                 if ($retryCounter -gt 0) {
-                    Write-Verbose -Message "Retrying in $RetryWaitTimeInSeconds seconds, $retryCounter attempts left"
-                    Start-Sleep -Seconds $RetryWaitTimeInSeconds
+                    $retryWaitTime = Get-WebSockerConnectWaitTime -Retry ($RetryCount - $retryCounter)
+                    Write-Verbose -Message "Retrying in $retryWaitTime seconds, $retryCounter attempts left"
+                    Start-Sleep -Seconds $retryWaitTime
                     continue
                 }
             }
@@ -116,7 +115,7 @@
 
             # Output connection object
             return [PSCustomObject]@{
-                URI                     = $URI
+                URI                     = $wssUri
                 WebSocket               = $webSocket
                 CancellationTokenSource = $cancellationTokenSource
                 RecvBuffer              = $recvBuffer